+++ sda/internal/database/postgres/method_deprecate_key_hash.go
@@ -3,7 +3,6 @@ package postgres
 import (
 	"context"
 	"database/sql"
-	"errors"
 )
 
 const deprecateKeyHashQuery = "deprecateKeyHash"
@@ -27,7 +26,7 @@ func (db *pgDb) deprecateKeyHash(ctx context.Context, tx *sql.Tx, keyHash string
 	}
 
 	if rowsAffected, _ := result.RowsAffected(); rowsAffected == 0 {
-		return errors.New("key hash not found or already deprecated")
+		return sql.ErrNoRows
 	}
 
 	return nil