+++ sda/CHANGELOG.md
@@ -10,6 +10,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - s3inbox: Allow forwarding of the [HeadObject action](https://docs.aws.amazon.com/AmazonS3/latest/API/API_HeadObject.html)
 - api: add new API /dataset/{dataset_id} for getting state of a dataset
+- Add observability package which adds support for tracing and metrics collecting and exporting with OpenTelemetry
+  - Add OpenTelemetry instrumentation for the PostgreSQL database connection
+  - Add support for injecting/extracting trace headers on the broker rabbitmq to support distributed tracing
+  - Add tracing in the storage reader/writer libraries for s3 and posix
+- Initialize observability package and otel middleware for http/grpc client and servers, and add span creating in the following
+  - api, auth, download, finalize, ingest, mapper, reencrypt, rotatekey, s3inbox, sync, and verify
 
 ### Changed